pipeline/requirements.txt
anthropic>=0.40.0
atproto>=0.0.54
Pillow>=10.0.0
requests>=2.32.0
python-dotenv>=1.0.0
openai>=1.0.0
